# docs/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
        "/api/auth/login": {
            "post": {
                "operationId": "AuthController_login",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns user profile and JWT token. is_default_password is true if the user has never changed their password."
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Invalid credentials or deactivated account"
                    },
                    "429": {
                        "description": "Too many login attempts \u2014 rate limit exceeded"
                    }
                },
                "summary": "Login and obtain a JWT access token",
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/auth/me": {
            "get": {
                "operationId": "AuthController_getMe",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Returns the current user profile"
                    },
                    "401": {
                        "description": "Missing or invalid token"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "summary": "Get current authenticated user",
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/auth/change-password": {
            "post": {
                "description": "Authenticated users change their own password. Sets is_default_password to false on success.",
                "operationId": "AuthController_changePassword",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePasswordDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "{ message: \"Password updated successfully\" }"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Current password is incorrect"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "summary": "Change own password",
                "tags": [
                    "Auth"
                ]
            }
        },
...
        "/api/alerts": {
            "get": {
                "description": "Returns only alerts belonging to the authenticated user, newest first.",
                "operationId": "AlertsController_findAll",
                "parameters": [
                    {
                        "name": "unread_only",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "type",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "enum": [
                                "deadline",
                                "missed_target",
                                "data_flag",
                                "sync_success",
                                "system"
                            ],
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "per_page",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plain array of alert objects (isRead, timestamp fields)"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "summary": "List alerts for the requesting user",
                "tags": [
                    "Alerts"
                ]
            }
        },
        "/api/alerts/readAll": {
            "post": {
                "operationId": "AlertsController_markAllRead",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "{ success: true }"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "summary": "Mark all alerts as read for the requesting user",
                "tags": [
                    "Alerts"
                ]
            }
        },
        "/api/alerts/{id}/read": {
            "post": {
                "operationId": "AlertsController_markRead",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Alert UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ success: true }"
                    },
                    "404": {
                        "description": "Alert not found"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "summary": "Mark a single alert as read",
                "tags": [
                    "Alerts"
                ]
            }
        },
...
    },
    "info": {
        "title": "EKZ M&E System API",
        "description": "Backend API for the Ekiti Knowledge Zone Monitoring & Evaluation System. Tracks programme performance through a logframe hierarchy, collects field data via mobile-friendly forms, and visualises progress through dashboards and GIS maps.",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "JWT": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http",
                "in": "header"
            }
        },
        "schemas": {
            "LoginDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "admin@ekz.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "Admin123!"
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
            "ChangePasswordDto": {
                "type": "object",
                "properties": {
                    "current_password": {
                        "type": "string",
                        "example": "Password12$"
                    },
                    "new_password": {
                        "type": "string",
                        "example": "NewSecure99!",
                        "description": "Min 8 characters. Must contain at least one uppercase letter, one number, and one special character."
                    }
                },
                "required": [
                    "current_password",
                    "new_password"
                ]
            },
...
        }
    }
}